# update sql queries for backend integration checks